+++ example/lib/browser_screen.dart
@@ -1,0 +1,21 @@
+import 'package:flutter/material.dart';
+import 'package:webview_flutter/webview_flutter.dart';
+
+class BrowserScreen extends StatelessWidget {
+  final String url;
+  final String title;
+
+  const BrowserScreen({super.key, required this.url, required this.title});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(title: Text(title)),
+      body: WebViewWidget(
+        controller: WebViewController()
+          ..setJavaScriptMode(JavaScriptMode.unrestricted)
+          ..loadRequest(Uri.parse(url)),
+      ),
+    );
+  }
+}