+++ example/lib/card_payment_tab.dart
@@ -1,0 +1,160 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_credit_card/flutter_credit_card.dart';
+import 'package:liqpay/liqpay.dart';
+import 'package:uuid/uuid.dart';
+
+class CardPaymentTab extends StatefulWidget {
+  final LiqPay liqPay;
+
+  const CardPaymentTab({super.key, required this.liqPay});
+
+  @override
+  State<CardPaymentTab> createState() => _CardPaymentTabState();
+}
+
+class _CardPaymentTabState extends State<CardPaymentTab> {
+  final GlobalKey<FormState> formKey = GlobalKey<FormState>();
+
+  late OutlineInputBorder border;
+
+  String _cardNumber = "";
+  String _cardExpiryDate = "";
+  String _cardHolderName = "";
+  String _cardCvvCode = "";
+  bool _isCvvFocused = false;
+
+  @override
+  void initState() {
+    super.initState();
+
+    border = OutlineInputBorder(
+      borderSide: BorderSide(
+        color: Colors.grey.withValues(alpha: 0.7),
+        width: 2.0,
+      ),
+    );
+
+    _cardNumber = "4242424242424242";
+    final now = DateTime.now();
+    _cardExpiryDate =
+        '${now.month.toString().padLeft(2, '0')}/${(now.year + 1).toString().substring(2)}';
+    _cardCvvCode = '000';
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.all(16),
+      child: Column(
+        mainAxisAlignment: MainAxisAlignment.center,
+        crossAxisAlignment: CrossAxisAlignment.stretch,
+        children: <Widget>[
+          CreditCardWidget(
+            enableFloatingCard: true,
+            glassmorphismConfig: null,
+            cardNumber: _cardNumber,
+            expiryDate: _cardExpiryDate,
+            cardHolderName: _cardHolderName,
+            cvvCode: _cardCvvCode,
+            bankName: 'Newgarden Solutions Bank',
+            showBackView: _isCvvFocused,
+            obscureCardNumber: true,
+            obscureCardCvv: true,
+            isHolderNameVisible: true,
+            cardBgColor: Colors.green,
+            backgroundImage: null,
+            isSwipeGestureEnabled: true,
+            onCreditCardWidgetChange: (CreditCardBrand creditCardBrand) {},
+          ),
+          Expanded(
+            child: SingleChildScrollView(
+              child: Column(
+                children: <Widget>[
+                  CreditCardForm(
+                    formKey: formKey,
+                    obscureCvv: true,
+                    obscureNumber: true,
+                    cardNumber: _cardNumber,
+                    cvvCode: _cardCvvCode,
+                    isHolderNameVisible: true,
+                    isCardNumberVisible: true,
+                    isExpiryDateVisible: true,
+                    cardHolderName: _cardHolderName,
+                    expiryDate: _cardExpiryDate,
+                    inputConfiguration: InputConfiguration(
+                      cardNumberDecoration: InputDecoration(
+                        labelText: 'Number',
+                        hintText: 'XXXX XXXX XXXX XXXX',
+                        hintStyle: const TextStyle(color: Colors.black),
+                        labelStyle: const TextStyle(color: Colors.black),
+                        focusedBorder: border,
+                        enabledBorder: border,
+                      ),
+                      expiryDateDecoration: InputDecoration(
+                        hintStyle: const TextStyle(color: Colors.black),
+                        labelStyle: const TextStyle(color: Colors.black),
+                        focusedBorder: border,
+                        enabledBorder: border,
+                        labelText: 'Expired Date',
+                        hintText: 'XX/XX',
+                      ),
+                      cvvCodeDecoration: InputDecoration(
+                        hintStyle: const TextStyle(color: Colors.black),
+                        labelStyle: const TextStyle(color: Colors.black),
+                        focusedBorder: border,
+                        enabledBorder: border,
+                        labelText: 'CVV',
+                        hintText: 'XXX',
+                      ),
+                      cardHolderDecoration: InputDecoration(
+                        hintStyle: const TextStyle(color: Colors.black),
+                        labelStyle: const TextStyle(color: Colors.black),
+                        focusedBorder: border,
+                        enabledBorder: border,
+                        labelText: 'Card Holder',
+                      ),
+                    ),
+                    onCreditCardModelChange: onCreditCardModelChange,
+                  ),
+                ],
+              ),
+            ),
+          ),
+          Padding(
+            padding: const EdgeInsets.all(16.0),
+            child: ElevatedButton(
+              onPressed: () async {
+                final cardDate = _cardExpiryDate.split('/');
+                final card = LiqPayCard(
+                  _cardNumber.trim(),
+                  cardDate[0],
+                  cardDate[1],
+                  _cardCvvCode,
+                );
+                final order = LiqPayOrder(
+                  const Uuid().v4(),
+                  1,
+                  'Test',
+                  card: card,
+                  action: LiqPayAction.auth,
+                );
+                await widget.liqPay.purchase(order);
+              },
+              child: const Text("PURCHASE"),
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+
+  void onCreditCardModelChange(CreditCardModel? creditCardModel) {
+    setState(() {
+      _cardNumber = creditCardModel!.cardNumber;
+      _cardExpiryDate = creditCardModel.expiryDate;
+      _cardHolderName = creditCardModel.cardHolderName;
+      _cardCvvCode = creditCardModel.cvvCode;
+      _isCvvFocused = creditCardModel.isCvvFocused;
+    });
+  }
+}